feat(YSP-288): create interface for modifier
